+++ animation_nodes/nodes/gpencil/gp_frame_info.py
@@ -30,7 +30,7 @@ class GPFrameInfoNode(bpy.types.Node, AnimationNode):
 
             self.newOutput(VectorizedSocket("GPStroke", "useIntegerList",
             ("Stroke", "stroke"), ("Strokes", "strokes")))
-        self.newOutput("Float", "Frame Number", "frameNumber")
+        self.newOutput("Integer", "Frame Number", "frameNumber")
 
     def draw(self, layout):
         layout.prop(self, "strokeType", text = "")
